rozdělení class detailu do jednotlivých tabů
