+++ frontend/src/classes/pages/ClassDetailPage.tsx
@@ -1,34 +1,6 @@
 import { useState, useEffect } from 'react';
 import { useParams, useNavigate } from 'react-router-dom';
-import {
-  Box,
-  Button,
-  Heading,
-  VStack,
-  Table,
-  Thead,
-  Tbody,
-  Tr,
-  Th,
-  Td,
-  Text,
-  Card,
-  CardHeader,
-  CardBody,
-  useToast,
-  Modal,
-  ModalOverlay,
-  ModalContent,
-  ModalHeader,
-  ModalFooter,
-  ModalBody,
-  ModalCloseButton,
-  useDisclosure,
-  Input,
-  Textarea,
-  TableContainer,
-  Stack,
-} from '@chakra-ui/react';
+import { Box, Button, useToast } from '@chakra-ui/react';
 import { ChevronLeftIcon } from '@chakra-ui/icons';
 import { texts } from '@frontend/texts';
 import { useLanguage } from '@frontend/shared/contexts/LanguageContext';
@@ -38,20 +10,11 @@ import { ROUTES } from '@frontend/shared/route';
 import { EditClassInfoModal } from '../components/EditClassInfoModal';
 import { ManageClassTeachersModal } from '../components/ManageClassTeachersModal';
 import Tabs, { TabItem } from '@frontend/shared/components/Tabs/Tabs';
+import InfoTab from '../tabs/InfoTab';
+import StudentsTab from '../tabs/StudentsTab';
+import ActivitiesTab from '../tabs/ActivitiesTab';
 
 import { Class } from '@frontend/types/class';
-
-interface ClassHistory {
-  id: number;
-  date: string;
-  notes: string;
-  created_at: string;
-  created_by: {
-    id: number;
-    firstname: string;
-    surname: string;
-  };
-}
 
 interface User {
   id: number;
@@ -74,11 +37,7 @@ const ClassDetailPage = () => {
   const { language } = useLanguage();
   const toast = useToast();
   const [classData, setClassData] = useState<Class | null>(null);
-  const [history, setHistory] = useState<ClassHistory[]>([]);
   const [nextActivities, setNextActivities] = useState<NextActivity[]>([]);
-  const { isOpen, onOpen, onClose } = useDisclosure();
-  const [newHistoryDate, setNewHistoryDate] = useState('');
-  const [newHistoryNotes, setNewHistoryNotes] = useState('');
   const [isEditInfoModalOpen, setIsEditInfoModalOpen] = useState(false);
   const [isMembersModalOpen, setIsMembersModalOpen] = useState(false);
   const [isAdmin, setIsAdmin] = useState(false);
@@ -115,21 +74,6 @@ const ClassDetailPage = () => {
           }
         };
 
-        const fetchHistory = async () => {
-          try {
-            const response = await api.get(`/api/classes/${id}/history`);
-            setHistory(response.data);
-          } catch (error) {
-            toast({
-              title: 'Error',
-              description: 'Failed to load class history',
-              status: 'error',
-              duration: 5000,
-              isClosable: true,
-            });
-          }
-        };
-
         const fetchNextActivities = async () => {
           try {
             if (!id) return;
@@ -148,7 +92,6 @@ const ClassDetailPage = () => {
 
         fetchClassData();
         if (id) {
-          fetchHistory();
           fetchNextActivities();
         }
 
@@ -169,43 +112,6 @@ const ClassDetailPage = () => {
 
     fetchData();
   }, [id, toast]);
-
-  const handleAddHistory = async () => {
-    try {
-      await api.post(`/api/classes/${id}/history`, {
-        date: newHistoryDate,
-        notes: newHistoryNotes,
-      });
-      const response = await api.get(`/api/classes/${id}/history`);
-      setHistory(response.data);
-      onClose();
-      setNewHistoryDate('');
-      setNewHistoryNotes('');
-    } catch (error) {
-      toast({
-        title: 'Error',
-        description: 'Failed to add history entry',
-        status: 'error',
-        duration: 5000,
-        isClosable: true,
-      });
-    }
-  };
-
-  const handleDeleteHistory = async (historyId: number) => {
-    try {
-      await api.delete(`/api/classes/${id}/history/${historyId}`);
-      setHistory(history.filter((h) => h.id !== historyId));
-    } catch (error) {
-      toast({
-        title: 'Error',
-        description: 'Failed to delete history entry',
-        status: 'error',
-        duration: 5000,
-        isClosable: true,
-      });
-    }
-  };
 
   const handleSaveClassInfo = async (updatedInfo: {
     name: string;
@@ -239,213 +145,39 @@ const ClassDetailPage = () => {
     }
   };
 
-  const getVisibleChildren = () => {
-    if (!classData) return [];
-    const acceptedChildren = classData.children.filter((child) => child.status === 'accepted');
-    if (isAdmin || isTeacher) return acceptedChildren;
-    if (isParent) {
-      return acceptedChildren.filter((child) => child.parent_id === currentUserId);
-    }
-    return [];
-  };
-
-  const getFilteredActivities = () => {
-    if (!classData) return [];
-    return nextActivities.filter((activity) => {
-      const child = classData.children.find((c) => c.id === activity.child_id);
-      return child && child.status === 'accepted';
-    });
-  };
-
   if (!classData) {
     return null;
   }
 
-  // Info Tab Content
   const infoTabContent = (
-    <VStack align="stretch" spacing={6}>
-      <Card>
-        <CardHeader>
-          <Heading size={{ base: 'sm', md: 'md' }}>{texts.classes.detail.info[language]}</Heading>
-        </CardHeader>
-        <CardBody>
-          <VStack align="stretch" spacing={4}>
-            <Box>
-              <Text fontWeight="bold">{texts.classes.name[language]}</Text>
-              <Text>{classData.name}</Text>
-            </Box>
-            <Box>
-              <Text fontWeight="bold">{texts.classes.description[language]}</Text>
-              <Text>{classData.description}</Text>
-            </Box>
-            <Box>
-              <Text fontWeight="bold">{texts.classes.ageRange[language]}</Text>
-              <Text>{`${classData.min_age} - ${classData.max_age}`}</Text>
-            </Box>
-            <Box>
-              <Text fontWeight="bold">{texts.classes.detail.teachers[language]}</Text>
-              {classData.teachers.map((teacher) => (
-                <Text key={teacher.id}>
-                  {teacher.firstname} {teacher.surname}
-                </Text>
-              ))}
-            </Box>
-          </VStack>
-          {isAdmin && (
-            <Stack mt={4} direction={{ base: 'column', sm: 'row' }} spacing={4} w="full">
-              <Button
-                colorScheme="blue"
-                onClick={() => setIsEditInfoModalOpen(true)}
-                size="md"
-                w={{ base: 'full', sm: 'auto' }}
-              >
-                {texts.classes.editInfo[language]}
-              </Button>
-              <Button
-                colorScheme="blue"
-                onClick={() => setIsMembersModalOpen(true)}
-                size="md"
-                w={{ base: 'full', sm: 'auto' }}
-              >
-                {texts.classes.teachers[language]}
-              </Button>
-            </Stack>
-          )}
-        </CardBody>
-      </Card>
-
-      <Card>
-        <CardHeader>
-          <Heading size={{ base: 'sm', md: 'md' }}>
-            {texts.classes.detail.history[language]}
-          </Heading>
-        </CardHeader>
-        <CardBody>
-          {(isAdmin || isTeacher) && (
-            <Button mb={4} colorScheme="blue" onClick={onOpen} size="md">
-              {texts.classes.detail.addHistory[language]}
-            </Button>
-          )}
-          <TableContainer>
-            <Table variant="simple" size="md">
-              <Thead>
-                <Tr>
-                  <Th>{texts.classes.detail.date[language]}</Th>
-                  <Th>{texts.classes.detail.notes[language]}</Th>
-                  <Th>{texts.classes.detail.createdBy[language]}</Th>
-                  {(isAdmin || isTeacher) && <Th>{texts.common.actions[language]}</Th>}
-                </Tr>
-              </Thead>
-              <Tbody>
-                {history.map((entry) => (
-                  <Tr key={entry.id}>
-                    <Td>
-                      {new Date(entry.date).toLocaleDateString(
-                        language === 'cs' ? 'cs-CZ' : 'en-US'
-                      )}
-                    </Td>
-                    <Td>{entry.notes}</Td>
-                    <Td>{`${entry.created_by.firstname} ${entry.created_by.surname}`}</Td>
-                    {(isAdmin || isTeacher) && (
-                      <Td>
-                        <Button
-                          size="sm"
-                          colorScheme="red"
-                          onClick={() => handleDeleteHistory(entry.id)}
-                        >
-                          {texts.common.delete[language]}
-                        </Button>
-                      </Td>
-                    )}
-                  </Tr>
-                ))}
-              </Tbody>
-            </Table>
-          </TableContainer>
-        </CardBody>
-      </Card>
-    </VStack>
+    <InfoTab
+      classData={classData}
+      language={language}
+      isAdmin={isAdmin}
+      onEditClick={() => setIsEditInfoModalOpen(true)}
+      onEditMembersClick={() => setIsMembersModalOpen(true)}
+    />
   );
 
   const studentsTabContent = (
-    <Card>
-      <CardHeader>
-        <Heading size={{ base: 'sm', md: 'md' }}>
-          {isParent
-            ? texts.classes.detail.myChildren[language]
-            : texts.classes.detail.students[language]}
-        </Heading>
-      </CardHeader>
-      <CardBody>
-        <TableContainer>
-          <Table variant="simple" size="md">
-            <Thead>
-              <Tr>
-                <Th>{texts.childrenTable.firstname[language]}</Th>
-                <Th>{texts.childrenTable.surname[language]}</Th>
-                <Th>{texts.childrenTable.age[language]}</Th>
-                {(isAdmin || isTeacher) && <Th>{texts.childrenTable.parent[language]}</Th>}
-                {(isAdmin || isTeacher) && <Th>{texts.childrenTable.contact[language]}</Th>}
-              </Tr>
-            </Thead>
-            <Tbody>
-              {getVisibleChildren().map((child) => (
-                <Tr key={child.id}>
-                  <Td>{child.firstname}</Td>
-                  <Td>{child.surname}</Td>
-                  <Td>{child.age}</Td>
-                  {(isAdmin || isTeacher) && <Td>{child.parent}</Td>}
-                  {(isAdmin || isTeacher) && <Td>{child.parent_contact || child.parent_email}</Td>}
-                  {(isAdmin || isTeacher) && (
-                    <Td>{child.status === 'pending' && classData.name}</Td>
-                  )}
-                </Tr>
-              ))}
-            </Tbody>
-          </Table>
-        </TableContainer>
-      </CardBody>
-    </Card>
+    <StudentsTab
+      classData={classData}
+      language={language}
+      isAdmin={isAdmin}
+      isTeacher={isTeacher}
+      isParent={isParent}
+      currentUserId={currentUserId}
+    />
   );
 
   const activitiesTabContent = (
-    <Card>
-      <CardHeader>
-        <Heading size={{ base: 'sm', md: 'md' }}>
-          {texts.classes.detail.nextActivities[language]}
-        </Heading>
-      </CardHeader>
-      <CardBody>
-        {getFilteredActivities().length === 0 ? (
-          <Text color="gray.500" fontStyle="italic">
-            {texts.classes.detail.noNextActivities[language]}
-          </Text>
-        ) : (
-          <TableContainer>
-            <Table variant="simple" size="md">
-              <Thead>
-                <Tr>
-                  <Th>{texts.childrenTable.firstname[language]}</Th>
-                  <Th>{texts.childrenTable.surname[language]}</Th>
-                  <Th>{texts.classes.detail.activity[language]}</Th>
-                  {(isAdmin || isTeacher) && <Th>{texts.classes.detail.notes[language]}</Th>}
-                </Tr>
-              </Thead>
-              <Tbody>
-                {getFilteredActivities().map((activity) => (
-                  <Tr key={`${activity.child_id}-${activity.id}`}>
-                    <Td>{activity.firstname}</Td>
-                    <Td>{activity.surname}</Td>
-                    <Td>{activity.activity || '-'}</Td>
-                    {(isAdmin || isTeacher) && <Td>{activity.notes || '-'}</Td>}
-                  </Tr>
-                ))}
-              </Tbody>
-            </Table>
-          </TableContainer>
-        )}
-      </CardBody>
-    </Card>
+    <ActivitiesTab
+      classData={classData}
+      nextActivities={nextActivities}
+      language={language}
+      isAdmin={isAdmin}
+      isTeacher={isTeacher}
+    />
   );
 
   const classDetailTabs: TabItem[] = [
@@ -478,36 +210,6 @@ const ClassDetailPage = () => {
       </Button>
 
       <Tabs tabs={classDetailTabs} variant="line" colorScheme="blue" />
-
-      {(isAdmin || isTeacher) && (
-        <Modal isOpen={isOpen} onClose={onClose} size={{ base: 'full', md: 'lg' }}>
-          <ModalOverlay />
-          <ModalContent>
-            <ModalHeader>{texts.classes.detail.addHistory[language]}</ModalHeader>
-            <ModalCloseButton />
-            <ModalBody>
-              <VStack spacing={4}>
-                <Input
-                  type="date"
-                  value={newHistoryDate}
-                  onChange={(e) => setNewHistoryDate(e.target.value)}
-                />
-                <Textarea
-                  value={newHistoryNotes}
-                  onChange={(e) => setNewHistoryNotes(e.target.value)}
-                  placeholder={texts.classes.detail.notesPlaceholder[language]}
-                />
-              </VStack>
-            </ModalBody>
-            <ModalFooter>
-              <Button colorScheme="blue" mr={3} onClick={handleAddHistory}>
-                {texts.common.save[language]}
-              </Button>
-              <Button onClick={onClose}>{texts.common.cancel[language]}</Button>
-            </ModalFooter>
-          </ModalContent>
-        </Modal>
-      )}
 
       {isAdmin && classData && (
         <>
